requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
boto==2.49.0
boto3==1.18.18
botocore==1.21.18
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
docutils==0.15.2
drf-yasg==1.20.0
ec2==0.4.0
et-xmlfile==1.1.0
Flask==2.0.1
futures==2.2.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==4.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multitasking==0.0.9
numpy==1.21.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.1
pandas-datareader==0.10.0
paramiko==2.7.2
plotly==5.1.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
tenacity==8.0.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
yfinance==0.1.63
